requirements.txt
reportlab==3.5.46
Pillow==7.2.0
PyPDF2==1.26.0
